Set XDG_CURRENT_DESKTOP to kde in runtime
Currently, QT fails to theme properly for GNOME (and potentially other desktops) inside of the Flatpak runtime. Forcefully setting XDG_CURRENT_DESKTOP to "kde" fixes this behavior.

A side-effect of this change is that the cursor will change its look from the host desktop while inside of the window, but this is preferable to the app's entire aspect looking outright broken.